requirements.txt
web3>=6.0.0
python-telegram-bot>=20.0
streamlit>=1.28.0
requests>=2.28.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
